MemberResDTO.GetHome numOfMissionSuccess 타입 Long으로 변경

+++ src/main/java/com/umc/storeandmission/domain/member/dto/MemberResDTO.java
@@ -12,8 +12,8 @@ public class MemberResDTO {
     ) {}
 
     public record GetHome(
-        Long point,
-        Integer numOfMissionSuccess
+            Long point,
+            Long numOfMissionSuccess
     ) {}
 
     public record GetCompleteCount(
